Bypass getFromFolder() error
In order to get this running, I need to comment out the code that finds the portraits on the server. When I figure out where a good way to do this, we can add portraits back in. For now though, we remove them, as the way this is supposed to work, doesn't appear to be working for me.